+++ lib/ex_doc/formatter/epub.ex
@@ -1,6 +1,6 @@
 defmodule ExDoc.Formatter.EPUB do
   @moduledoc """
-  Provide EPUB documentation
+  Generates EPUB documentation for Elixir projects.
   """
 
   @mimetype "application/epub+zip"
@@ -8,7 +8,7 @@ defmodule ExDoc.Formatter.EPUB do
   alias ExDoc.Formatter.HTML
 
   @doc """
-  Generate EPUB documentation for the given modules
+  Generate EPUB documentation for the given modules.
   """
   @spec run(list, ExDoc.Config.t) :: String.t
   def run(project_nodes, config) when is_map(config) do
@@ -26,7 +26,6 @@ defmodule ExDoc.Formatter.EPUB do
     nodes_map = %{
       modules: HTML.filter_list(:module, all),
       exceptions: HTML.filter_list(:exception, all),
-      protocols: HTML.filter_list(:protocol, all),
       tasks: HTML.filter_list(:task, all)
     }
 
@@ -39,7 +38,6 @@ defmodule ExDoc.Formatter.EPUB do
     generate_extras(config)
     generate_list(config, nodes_map.modules)
     generate_list(config, nodes_map.exceptions)
-    generate_list(config, nodes_map.protocols)
     generate_list(config, nodes_map.tasks)
 
     {:ok, epub} = generate_epub(config.output)
@@ -51,7 +49,7 @@ defmodule ExDoc.Formatter.EPUB do
     output =
       config.output
       |> Path.expand()
-      |> Path.join("#{config.project}-v#{config.version}")
+      |> Path.join("#{config.project}")
     %{config | output: output}
   end