+++ lib/ex_doc/formatter/epub/assets.ex
@@ -1,10 +1,8 @@
 defmodule ExDoc.Formatter.EPUB.Assets do
   @moduledoc false
 
-  @assets_dir Application.app_dir(:ex_doc, "priv/ex_doc/formatter/epub/assets")
-
   defmacrop embed_pattern(pattern) do
-    [@assets_dir, pattern]
+    ["formatters/epub", pattern]
     |> Path.join()
     |> Path.wildcard()
     |> Enum.map(&{Path.basename(&1), File.read!(&1)})