+++ test/ex_doc/formatter/html/templates_test.exs
@@ -14,8 +14,8 @@ defmodule ExDoc.Formatter.HTML.TemplatesTest do
     "http://elixir-lang.org"
   end
 
-  defp doc_config do
-    %ExDoc.Config{
+  defp doc_config(config \\ []) do
+    default = %ExDoc.Config{
       project: "Elixir",
       version: "1.0.1",
       source_root: File.cwd!,
@@ -24,6 +24,8 @@ defmodule ExDoc.Formatter.HTML.TemplatesTest do
       source_url: source_url(),
       output: "test/tmp/html_templates"
     }
+
+    struct(default, config)
   end
 
   defp get_module_page(names) do
@@ -37,7 +39,7 @@ defmodule ExDoc.Formatter.HTML.TemplatesTest do
 
   setup_all do
     File.mkdir_p!("test/tmp/html_templates")
-    File.cp_r!("priv/ex_doc/formatter/html/assets", "test/tmp/html_templates")
+    File.cp_r!("formatters/html", "test/tmp/html_templates")
     File.touch!("test/tmp/html_templates/dist/sidebar_items-123456.js")
     :ok
   end
@@ -209,6 +211,15 @@ defmodule ExDoc.Formatter.HTML.TemplatesTest do
     assert content =~ ~r("id":"CompiledWithDocs".*"functions":.*"example/2")ms
     assert content =~ ~r("id":"CompiledWithDocs".*"functions":.*"example_without_docs/0")ms
     assert content =~ ~r("id":"CompiledWithDocs.Nested")ms
+  end
+
+  test "list_page outputs groups for the given nodes" do
+    names = [CompiledWithDocs, CompiledWithDocs.Nested]
+    group_mapping = [groups_for_modules: ["Group": [CompiledWithDocs]]]
+    nodes = ExDoc.Retriever.docs_from_modules(names, doc_config(group_mapping))
+    content = Templates.create_sidebar_items(%{modules: nodes}, [])
+
+    assert content =~ ~r("id":"CompiledWithDocs","title":"CompiledWithDocs","group":"Group")ms
   end
 
   ## MODULES